pshishing-detection-api/requirements.txt
Flask==2.1.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.5
filelock==3.16.1
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
threadpoolctl==3.5.0
tldextract==5.1.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whois==1.20240129.2
wsproto==1.2.0
xgboost==2.1.2
